requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
pytest>=7.0.0
